ci(test): cache leetcode db test: tweak
