+++ .github/workflows/release.yml
@@ -24,6 +24,7 @@ jobs:
         uses: dtolnay/rust-toolchain@master
         with:
           toolchain: nightly
+
       - name: Install cargo-binstall
         uses: taiki-e/install-action@v2
         with:
