+++ crates/leetcode-api/tests/leetcode_work.rs
@@ -1,47 +1,18 @@
-use leetcode_api::{
-    glob_leetcode,
-    leetcode::{question::qs_detail::Question, IdSlug},
-    render::Render,
-};
+use leetcode_api::{glob_leetcode, leetcode::IdSlug};
 use miette::Result;
 use pretty_assertions::assert_eq;
 
-#[ignore = "manual"]
-#[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-async fn get_all_pbs_works() -> Result<()> {
-    // tracing_subscriber::fmt()
-    //     .with_max_level(tracing::Level::DEBUG)
-    //     .with_test_writer()
-    //     .init();
-
-    glob_leetcode()
-        .await
-        .sync_problem_index()
-        .await?;
-    Ok(())
-}
-
-#[ignore = "manual"]
-#[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-async fn new_get_index() -> Result<()> {
-    // tracing_subscriber::fmt()
-    //     .with_max_level(tracing::Level::DEBUG)
-    //     .with_test_writer()
-    //     .init();
-
-    glob_leetcode()
-        .await
-        .sync_index_topic()
-        .await?;
-    Ok(())
-}
-
-#[ignore]
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
 async fn get_user_code_work() -> Result<()> {
+    let id = IdSlug::Id(108);
+    glob_leetcode()
+        .await
+        .get_qs_detail(id.clone(), false)
+        .await?;
+
     let a = glob_leetcode()
         .await
-        .get_user_code(IdSlug::Id(108))
+        .get_user_code(id)
         .await?;
 
     assert!(!a.0.is_empty());
@@ -50,63 +21,15 @@ async fn get_user_code_work() -> Result<()> {
     Ok(())
 }
 
-#[ignore = "manual"]
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-async fn test_work() -> Result<()> {
-    tracing_subscriber::fmt()
-        .with_max_level(tracing::Level::DEBUG)
-        .with_test_writer()
-        .init();
-
-    match glob_leetcode()
+async fn get_qs_detail_none() {
+    assert!(glob_leetcode()
         .await
-        .test_code(IdSlug::Id(435))
+        .get_qs_detail(IdSlug::Id(0), false)
         .await
-    {
-        Ok((_, res)) => {
-            dbg!(&res);
-            println!(r##"(| res |) -> {} "##, res.to_md_str(false));
-            // res.render_with_mdcat();
-            // dbg!(res.to_tui_vec());
-        },
-        Err(e) => eprintln!("{e}"),
-    }
-    match glob_leetcode()
-        .await
-        .test_code(IdSlug::Id(435))
-        .await
-    {
-        Ok((_, res)) => {
-            dbg!(&res);
-            println!(r##"(| res |) -> {} "##, res.to_md_str(false));
-            // res.render_with_mdcat();
-            // dbg!(res.to_tui_vec());
-        },
-        Err(e) => eprintln!("{e}"),
-    }
-
-    Ok(())
+        .is_err());
 }
 
-#[ignore = "need realy environment"]
-#[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-async fn submit_work() {
-    // tracing_subscriber::fmt()
-    //     .with_max_level(tracing::Level::DEBUG)
-    //     .with_test_writer()
-    //     .init();
-
-    let (_, res) = glob_leetcode()
-        .await
-        .submit_code(IdSlug::Id(27))
-        .await
-        .unwrap();
-    dbg!(res.to_tui_vec());
-    println!(r##"(| res |) -> {} "##, res.to_md_str(false));
-    res.render_with_mdcat();
-}
-
-#[ignore]
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
 async fn get_qs_detail_work() -> Result<()> {
     // tracing_subscriber::fmt()
@@ -122,10 +45,6 @@ async fn get_qs_detail_work() -> Result<()> {
         )
         .await
         .unwrap();
-    // println!("{:#?}", question.meta_data);
-    // println!("{:#?}", question.stats);
-    // println!("{:#?}", question.env_info);
-    // dbg!(&question);
     assert_eq!(
         &question.qs_slug.unwrap(),
         "find-smallest-common-element-in-all-rows"
@@ -138,10 +57,6 @@ async fn get_qs_detail_work() -> Result<()> {
         &question.sample_test_case,
         "[[1,2,3,4,5],[2,4,5,8,10],[3,5,7,9,11],[1,3,5,7,9]]"
     );
-    assert_eq!(
-        &question.translated_title.unwrap(),
-        "找出所有行中最小公共元素"
-    );
     assert_eq!(&question.question_id, "1143");
     assert_eq!(
         &question.question_title.unwrap(),
@@ -153,7 +68,6 @@ async fn get_qs_detail_work() -> Result<()> {
         .get_qs_detail(IdSlug::Slug("two-sum".to_owned()), true)
         .await?;
     assert_eq!(&question.question_id, "1");
-    assert_eq!(&question.translated_title.unwrap(), "两数之和");
     assert_eq!(&question.title, "Two Sum");
     assert_eq!(&question.qs_slug.unwrap(), "two-sum");
     assert_eq!(&question.question_title.unwrap(), "Two Sum");
@@ -177,94 +91,6 @@ async fn get_qs_detail_work() -> Result<()> {
     assert_eq!(&question.question_id, "195");
     assert_eq!(&question.question_title.unwrap(), "Tenth Line");
     assert_eq!(&question.title, "Tenth Line");
-
-    Ok(())
-}
-
-#[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-async fn get_qs_detail_none() {
-    let question = glob_leetcode()
-        .await
-        .get_qs_detail(IdSlug::Id(0), false)
-        .await
-        .unwrap();
-    assert_eq!(question, Question::default());
-}
-
-#[ignore = "manual"]
-#[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-async fn get_submit_list() -> Result<()> {
-    // tracing_subscriber::fmt()
-    //     .with_max_level(tracing::Level::DEBUG)
-    //     .with_test_writer()
-    //     .init();
-
-    let a = glob_leetcode().await;
-    let res = a.all_submit_res(IdSlug::Id(32)).await?;
-    println!("{}", res);
-    // render_str(res.to_string())?;
-    // let res = get_rendered_str(res.to_string(), 30, 10)?;
-    // println!("{}", res);
-
-    Ok(())
-}
-#[ignore = "manual"]
-#[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-async fn daily_checkin() -> Result<()> {
-    // tracing_subscriber::fmt()
-    //     .with_max_level(tracing::Level::DEBUG)
-    //     .with_test_writer()
-    //     .init();
-
-    let a = glob_leetcode().await;
-    let res = a.daily_checkin().await;
-    println!("{:#?}", res);
-
-    Ok(())
-}
-#[ignore = "manual"]
-#[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-async fn user_points() -> Result<()> {
-    // tracing_subscriber::fmt()
-    //     .with_max_level(tracing::Level::DEBUG)
-    //     .with_test_writer()
-    //     .init();
-
-    let a = glob_leetcode().await;
-    let res = a.get_points().await?;
-    println!("{:#?}", res);
-
-    Ok(())
-}
-#[ignore = "manual"]
-#[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-async fn user_global_data() -> Result<()> {
-    // tracing_subscriber::fmt()
-    //     .with_max_level(tracing::Level::DEBUG)
-    //     .with_test_writer()
-    //     .init();
-
-    let a = glob_leetcode().await;
-    let res = a.get_user_info().await?;
-    println!("{:#?}", res);
-
-    Ok(())
-}
-
-#[ignore = "manual"]
-#[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-async fn user_pass_data() -> Result<()> {
-    // tracing_subscriber::fmt()
-    //     .with_max_level(tracing::Level::DEBUG)
-    //     .with_test_writer()
-    //     .init();
-
-    let a = glob_leetcode().await;
-    let temp = a.get_user_info().await?;
-    let res = a
-        .pass_qs_status(&temp.user_slug.unwrap_or_default())
-        .await?;
-    println!("{:#?}", res);
 
     Ok(())
 }
