+++ crates/leetcode-api/tests/query_work.rs
@@ -1,7 +1,6 @@
 use leetcode_api::{dao::query::*, entities::topic_tags};
 use miette::Result;
 
-#[ignore]
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
 async fn query_base() -> Result<()> {
     // tracing_subscriber::fmt()
@@ -23,7 +22,6 @@ async fn query_base() -> Result<()> {
     Ok(())
 }
 
-#[ignore]
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
 async fn query_count() -> Result<()> {
     let a = Query::query_status().await?;
@@ -35,7 +33,6 @@ async fn query_count() -> Result<()> {
     Ok(())
 }
 
-#[ignore]
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
 async fn query_all_topic_tags() -> Result<()> {
     let alltop: Vec<topic_tags::Model> = Query::query_all_topic().await?;
